organized E2E test files
