+++ e2e/protractor.conf.js
@@ -5,7 +5,8 @@ const { SpecReporter } = require("jasmine-spec-reporter");
 
 exports.config = {
   allScriptsTimeout: 11000,
-  specs: ["./src/specs/*.e2e-spec.ts"],
+  // specs: ["./src/specs/*e2e-spec.ts"],
+  specs:["./src/Features/Monitoring/CreateMonitor/*e2e-spec.ts"],
   capabilities: {
     browserName: "chrome",
     chromeOptions: {
