+++ e2e/src/features/misc/BreadTest.e2e-spec.ts
@@ -1,6 +1,6 @@
-import { AppPage } from "../pageObjects/app.po";
-import { navigations } from "../commons/navigations";
+
 import { browser, element, by } from "protractor";
+import { AppPage } from "e2e/src/pages/app.po";
 
 describe("Breadcrumb testing in resources", () => {
   let page: AppPage;
