+++ e2e/src/features/misc/app.e2e-spec.ts
@@ -1,4 +1,5 @@
-import { AppPage } from "../pageObjects/app.po";
+import { AppPage } from "e2e/src/pages/app.po";
+
 
 describe("workspace-project App", () => {
   let page: AppPage;
