+++ e2e/src/features/misc/navigations.e2e-spec.ts
@@ -1,12 +1,13 @@
-import { navigations } from "../commons/navigations";
+
 import { browser } from "protractor";
-import { AppPage } from "../pageObjects/app.po";
 import { async } from "@angular/core/testing";
+import { AppPage } from "e2e/src/pages/app.po";
+import { navigations } from "e2e/src/commons/navigations";
 
 
 describe("App Navigations", () => {
   let nav: navigations;
-  let page: AppPage;
+  let page:AppPage;
   
 
   beforeEach(() => {
