+++ e2e/src/features/misc/pagination.e2e-spec.ts
@@ -1,6 +1,6 @@
 import { browser, element, by } from "protractor";
-import { AppPage } from "../pageObjects/app.po";
 import { async } from "@angular/core/testing";
+import { AppPage } from "e2e/src/pages/app.po";
