+++ e2e/src/features/monitoring/createmonitor/portTest.e2e-spec.ts
@@ -1,5 +1,5 @@
-import { AppPage } from "../pageObjects/app.po";
-import { navigations } from "../commons/navigations";
+import { AppPage } from "../../../pages/app.po";
+import { navigations } from "../../../commons/navigations";
 import { browser, element, by } from "protractor";
 
 describe("Min Max Port field create monitor", () => {
