+++ e2e/src/features/resources/resourcedetails/labels_resources.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { AppPage } from "../pageObjects/app.po";
+import { AppPage } from "../../../pages/app.po";
 import { element, by, browser, Browser } from "protractor";
 
 describe("Add-fields Component Test in resources", () => {
